Fix system tests
